# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Grounded question answering over internal policy documents.",
    "title": "Sourcebook API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/documents": {
      "get": {
        "description": "Paginated, searchable list of indexed documents.\n\nq        \u2014 case-insensitive substring match on title\ncategory \u2014 exact category filter\nlimit    \u2014 page size (default 50, capped at 200)\nskip     \u2014 pagination offset",
        "operationId": "list_documents_api_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "default": "",
              "title": "Category",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents"
      }
    },
    "/api/documents/body": {
      "get": {
        "description": "The full markdown body of one document \u2014 powers the Policy Library reader.\n\nServed from the reading copy ingestion stores per document, never stitched\nfrom passages: chunks overlap, so joining them would repeat text. Two 404s:\na source the library does not know, and one it lists whose body was never\nstored because the corpus was ingested before bodies were kept. The client\ntreats both the same, falling back to passages, but the detail says which.",
        "operationId": "get_document_body_api_documents_body_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document Body"
      }
    },
    "/api/documents/categories": {
      "get": {
        "description": "Distinct categories present in the corpus, for filter controls.",
        "operationId": "list_categories_api_documents_categories_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Categories"
      }
    },
    "/api/documents/passages": {
      "get": {
        "description": "Ordered passages for one document \u2014 the citation audit trail.\n\nShows the exact text the assistant retrieves from, not a re-rendering of\nthe original file. The source pane beside an answer uses this; the Policy\nLibrary reader uses /documents/body.",
        "operationId": "get_document_passages_api_documents_passages_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document Passages"
      }
    },
    "/api/documents/reindex": {
      "post": {
        "description": "Force a rebuild of the document library and invalidate cached answers.\n\nBumping the corpus version is what clears the answer cache \u2014 entries are\nkeyed by it, so old ones simply stop matching.",
        "operationId": "reindex_documents_api_documents_reindex_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reindex Documents"
      }
    },
...
  }
}